requirements.txt
pyTelegramBotAPI
pydantic_settings
python-dotenv